added address controller and services sample

+++ BlogApi/Models/Entities/AddressEntity.cs
@@ -4,9 +4,9 @@ namespace BlogApi.Entity;
 
 public class AddressEntity
 {
-    public Int64 objectId { get; set; }
+    public Int32 objectId { get; set; }
     public Guid objectGuid { get; set; }
-    public Int64 parentId { get; set; }
+    public Int32 parentId { get; set; }
     public String text { get; set; }
     public GarAddressLevel objectLevel { get; set; }
 }
